[ADD] str functions reformat

+++ lib/my/my_str/c_string/my_strdup.c
@@ -10,7 +10,7 @@
 
 char *my_strdup(char const *src)
 {
-    int length = my_strlen(src);
+    size_t length = my_strlen(src);
     char *dest = malloc(sizeof(char) * (length + 1));
 
     if (dest == NULL)
